# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Warn:
    properties:
      guildID:
        type: string
      memberID:
        type: string
      warnID:
        type: integer
      warnReason:
        type: string
      warnedAt:
        type: string
      warnerID:
        type: string
    type: object
host: cardinal.gyroskan.com
info:
  contact:
    email: gyroskan@gmail.com
    name: API Support
  description: The API to interact with cardinal discord bot database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Cardinal API
  version: "0.1"
paths:
...
  /guilds/{guildID}/members/{memberID}/warns:
    get:
      description: Fetch all warns of the member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Member Warns
      tags:
      - Warns
    post:
      consumes:
      - application/json
      description: Create a new warn for a member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created warn
          schema:
            $ref: '#/definitions/models.Warn'
        "400":
          description: Wrong values
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Create warn
      tags:
      - Warns
  /guilds/{guildID}/members/{memberID}/warns/{warnID}:
    delete:
      consumes:
      - application/json
      description: Delete a member's warn
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      - description: warn id
        in: path
        name: warnID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server Error
      summary: Delete member's warn
      tags:
      - Warns
    get:
      description: Fetch the warn of the member.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      - description: warn id
        in: path
        name: warnID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one warn
      tags:
      - Warns
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"